+++ changelog.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Author in API: View all, create and delete
 - Cover in API: View all, create and delete
+- Notify in API: View all, create and delete
 
 ## [0.6.0] - 2020-12-07
 ### Added